requirements.dev.txt
-r requirements.txt
pytest-asyncio
pytest
pytest-cov
pytest-randomly
types-requests
asyncpg-stubs==0.29.0
